requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-vote==2.1.7
djangorestframework==3.11.0
greencache==1.0.0
htmq==1.0.0
idna==2.9
Pillow==7.1.1
psycopg2==2.8.5
pytz==2019.3
redcmd==1.2.10
redlib==1.5.7
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
vote==1.0.2
zope.interface==5.1.0
